systemctl: add systemctl show-log command
Add new unit command `show-log` to systemctl for more fine grained mac.

Currently the way to access logs for a unit is via `journalctl -u unitname`.
Journalctl reads the journal from disk and filters out the relevant messages.
This patch adds the command `systemctl show-log unitname`.

On a SELinux enabled system, where root might be constrained by the SELinux context, the current way does not offer the possibility to constrain the access to only specific unit logs based on the SELinux context.
`journalctl -u unitname` needs read permissions on the journal and so the SELinux context needs the permit to access these files. But with this read permission, one could real all logs.

With this new method, one can control the access of the SELinux context regarding what unit he can read logs for.
It is like the "start" permission: the SELinux policy needs to allow a context to "start" a specific unit by the unit label. Now the policy can also define which root context can read which unit logs by the permission  "showlog" on the class "service" with the target context of the unit label.